+++ src/components/app/details/cdDetails/cd.type.ts
@@ -9,35 +9,12 @@ export interface DeploymentHistoryDetail {
     values: Record<string, DeploymentHistorySingleValue>
     codeEditorValue: DeploymentHistorySingleValue
 }
-export interface DeploymentTemplateHistoryType {
-    currentConfiguration: DeploymentHistoryDetail
-    baseTemplateConfiguration: DeploymentHistoryDetail
-    previousConfigAvailable: boolean
-}
-export interface DeploymentTemplateOptions {
-    label: string
-    value: string
-    author: string
-    status: string
-}
-export interface CompareWithBaseConfiguration {
-    selectedDeploymentTemplate: DeploymentTemplateOptions
-    setSelectedDeploymentTemplate: (selected) => void
-    setFullScreenView: React.Dispatch<React.SetStateAction<boolean>>
-    setLoader: React.Dispatch<React.SetStateAction<boolean>>
-    setPreviousConfigAvailable: React.Dispatch<React.SetStateAction<boolean>>
-}
 export interface HistoryDiffSelectorList {
     id: number
     deployedOn: string
     deployedBy: string
     deploymentStatus: string
     wfrId?: number
-}
-export interface CompareViewDeploymentType {
-    setFullScreenView: React.Dispatch<React.SetStateAction<boolean>>
-    deploymentHistoryList: DeploymentTemplateList[]
-    setDeploymentHistoryList: React.Dispatch<React.SetStateAction<DeploymentTemplateList[]>>
 }
 export interface DeploymentTemplateList {
     id: number
@@ -61,21 +38,6 @@ export interface DeploymentHistory {
     email_id?: string
     image: string
     workflow_type?: string
-}
-
-export interface DeploymentHistoryParamsType {
-    appId: string
-    pipelineId?: string
-    historyComponent?: string
-    baseConfigurationId?: string
-    historyComponentName?: string
-    envId?: string
-    triggerId?: string
-}
-
-export interface DeploymentHistorySidebarType {
-    deploymentHistoryList: DeploymentTemplateList[]
-    setDeploymentHistoryList: React.Dispatch<React.SetStateAction<DeploymentTemplateList[]>>
 }
 
 export interface DeploymentDetailStepsType{
