+++ src/components/app/details/cdDetails/deploymentHistoryDiff/DeploymentHistoryConfigList.component.tsx
@@ -2,7 +2,8 @@ import React, { useEffect, useState } from 'react'
 import { ReactComponent as RightArrow } from '../../../../../assets/icons/ic-arrow-left.svg'
 import { NavLink } from 'react-router-dom'
 import { useRouteMatch, useParams } from 'react-router'
-import { DeploymentHistoryParamsType, DeploymentTemplateList } from '../cd.type'
+import { DeploymentTemplateList } from '../cd.type'
+import { DeploymentHistoryParamsType } from './types'
 import { getDeploymentHistoryList } from '../service'
 import { DEPLOYMENT_HISTORY_CONFIGURATION_LIST_MAP } from '../../../../../config'
 import CDEmptyState from '../CDEmptyState'
