+++ src/components/app/details/cdDetails/deploymentHistoryDiff/DeploymentHistoryHeader.tsx
@@ -5,7 +5,7 @@ import { NavLink } from 'react-router-dom'
 import moment from 'moment'
 import { DEPLOYMENT_HISTORY_CONFIGURATION_LIST_MAP, Moment12HourFormat, URLS } from '../../../../../config'
 import { ReactComponent as LeftIcon } from '../../../../../assets/icons/ic-arrow-forward.svg'
-import { CompareWithBaseConfiguration, DeploymentHistoryParamsType, DeploymentTemplateOptions } from '../cd.type'
+import { CompareWithBaseConfiguration, DeploymentHistoryParamsType, DeploymentTemplateOptions } from './types'
 import { Option, styles } from './utils'
 import { getDeploymentDiffSelector } from '../service'
 import { showError } from '../../../../common'
